pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Web application for checking a web page"
authors = ["piranha887 <ilyanikitin889@gmail.com>"]
repository = "https://github.com/piranha887/python-project-83.git"
classifiers = [
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
gunicorn = "20.1.0"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.11.2"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"